requirements.txt
Django==4.1.7
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
PyJWT==2.6.0
pytest==7.2.1
pytest-django==4.5.2